+++ stat/util/ip_locator.go
@@ -11,33 +11,25 @@ import (
 	"github.com/KyberNetwork/reserve-data/common"
 )
 
-// IPToCountry initializes a new ipLocator and resolve the given IP address.
-func IPToCountry(ip string) (string, error) {
-	dbPath := path.Join(common.CurrentDir(), "GeoLite2-Country.mmdb")
-	il, err := newIPLocator(dbPath)
-	if err != nil {
-		// MaxMind's DB is stored in source tree
-		panic(err)
-	}
-	return il.ipToCountry(ip)
-}
+const geoDBFile = "GeoLite2-Country.mmdb"
 
 // ipLocator is a resolver that query data of IP from MaxMind's GeoLite2 database.
-type ipLocator struct {
+type IPLocator struct {
 	r *geoip2.Reader
 }
 
 // newIPLocator returns an instance of ipLocator.
-func newIPLocator(dbPath string) (*ipLocator, error) {
+func NewIPLocator() (*IPLocator, error) {
+	dbPath := path.Join(common.CurrentDir(), geoDBFile)
 	r, err := geoip2.Open(dbPath)
 	if err != nil {
 		return nil, err
 	}
-	return &ipLocator{r: r}, nil
+	return &IPLocator{r: r}, nil
 }
 
-// ipToCountry returns the country of given IP address.
-func (il *ipLocator) ipToCountry(ip string) (string, error) {
+// IPToCountry returns the country of given IP address.
+func (il *IPLocator) IPToCountry(ip string) (string, error) {
 	IPParsed := net.ParseIP(ip)
 	if IPParsed == nil {
 		return "", fmt.Errorf("ip %s is not valid!", ip)